+++ CMakeLists.txt
@@ -6,6 +6,7 @@ add_subdirectory(errors)
 add_subdirectory(atlas)
 add_subdirectory(temperature)
 add_subdirectory(mcp3008)
+add_subdirectory(configuration)
 add_subdirectory(app)
 add_subdirectory(calibrate_ph)
 add_subdirectory(calibrate_do)
