+++ app/CMakeLists.txt
@@ -18,11 +18,13 @@ include_directories (${CMAKE_SOURCE_DIR}/timer
                     ${CMAKE_SOURCE_DIR}/mcp3008 
                     ${CMAKE_SOURCE_DIR}/temperature 
                     ${CMAKE_SOURCE_DIR}/errors 
-                    ${CMAKE_SOURCE_DIR}/mqtt)
+                    ${CMAKE_SOURCE_DIR}/mqtt
+                    ${CMAKE_SOURCE_DIR}/configuration)
                     
 add_executable (${PROJECT_NAME} ${SOURCES})
 
-target_link_libraries (${PROJECT_NAME} Threads::Threads -lwiringPi -lconfig++ -lpaho-mqttpp3 -lpaho-mqtt3as
+target_link_libraries (${PROJECT_NAME} ${COMMON_FLAGS} Threads::Threads -lwiringPi -lconfig++ -lpaho-mqttpp3 -lpaho-mqtt3as
+                    ${CMAKE_BINARY_DIR}/configuration/libconfiguration.a
                     ${CMAKE_BINARY_DIR}/atlas/libatlas.a 
                     ${CMAKE_BINARY_DIR}/timer/libtimer.a 
                     ${CMAKE_BINARY_DIR}/mcp3008/libmcp3008.a
